microservices/synthesis/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
azure-core==1.39.0
azure-storage-blob==12.28.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
distro==1.9.0
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
isodate==0.7.2
jiter==0.14.0
numpy==2.4.4
openai==2.32.0
pandas==3.0.2
pyarrow==24.0.0
pycparser==3.0
pydantic==2.13.3
pydantic-core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.46.0
